PersonalBlog-Flask/requirements.txt
astroid==2.4.2
click==7.1.2
colorama==0.4.3
decorator==4.4.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pbr==5.4.5
pylint==2.5.3
six==1.15.0
SQLAlchemy==1.3.18
sqlalchemy-migrate==0.13.0
sqlparse==0.3.1
Tempita==0.5.2
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
